+++ modules/node-groups/modules/node-group/local.tf
@@ -3,14 +3,12 @@ locals {
 
   os_types = {
     "ubuntu"      = "Linux"
-    "windows"     = "Windows"
     "windows2019" = "Windows"
     "windows2022" = "Windows"
   }
 
   os_skus = {
     "ubuntu"      = "Ubuntu"
-    "windows"     = null
     "windows2019" = "Windows2019"
     "windows2022" = "Windows2022"
   }